+++ src/main/java/com/bodimTikka/bodimTikka/dto/UserProjection.java
@@ -1,4 +1,4 @@
-package com.bodimTikka.bodimTikka.DTO;
+package com.bodimTikka.bodimTikka.dto;
 
 public interface UserProjection {
     Long getId();
