+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.application.name=bodimtikka
 
-spring.datasource.url=jdbc:mysql://${DB_HOST}:${DB_PORT}/${DB_NAME}?sslmode=require
+spring.datasource.url=jdbc:postgresql://${DB_HOST}:${DB_PORT}/${DB_NAME}?sslmode=require
 spring.datasource.username=${DB_USERNAME}
 spring.datasource.password=${DB_PASSWORD}
 spring.datasource.driver-class-name=org.postgresql.Driver
@@ -8,3 +8,4 @@ spring.datasource.driver-class-name=org.postgresql.Driver
 # Hibernate (JPA) settings
 #spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
+server.port=9090
